requirements.txt
django==4.1.7
httpx==0.23.3
requests==2.28.2
pyairtable==1.4.0
pytz==2022.7.1
djangorestframework==3.14.0
django-redis==5.2.0
python-decouple==3.7
celery==5.2.7
django-celery-beat==2.5.0
redis==4.5.2
kombu==5.2.4
dj-database-url==1.2.0
psycopg2==2.9.5
django-bulk-update-or-create==0.3.0
playwright==1.31.1
gunicorn==20.1.0
python-dateutil~=2.8.2
pymongo==4.3.3
slack-sdk==3.20.2
